History Items : Save/Unsave behavior corrected
A browsed FeedObject gets saved to FEED_TABLE by default, as hidden.
Then when the user chooses to Save from History, this DB entry becomes visible.

Every browsed FeedObject stays in the database indefinitely this way.

Note that the alternative would be not recording FeedObject to FEED_TABLE upon browsing, by default.
In this alternative, a request to watrcoolr.js would be required to read FeedObject details again, while saving.